requirements.txt
pyodbc
python-dotenv
pytelegrambotapi 